+++ src/core/ids-attributes.ts
@@ -410,6 +410,7 @@ export const attributes = {
   SORTABLE: 'sortable',
   SOURCE_FORMATTER: 'source-formatter',
   SQUARE: 'square',
+  SELF_MANAGED: 'self-managed',
   SRC: 'src',
   STACKED: 'stacked',
   STARS: 'stars',